backend/requirements.txt
fastapi
uvicorn
litellm
python-dotenv
chromadb
pypdf
pytest
 